heatmap/requirements.txt
appdirs==1.4.0
cycler==0.10.0
matplotlib==2.0.0
numpy==1.12.0
packaging==16.8
pyparsing==2.1.10
pyserial==3.2.1
python-dateutil==2.6.0
pytz==2016.10
six==1.10.0
